TASK: Remove new non-PSR-7 methods from Request
