+++ TYPO3.Flow/Tests/Unit/Http/RequestTest.php
@@ -408,51 +408,6 @@ class RequestTest extends UnitTestCase
     /**
      * Data Provider
      */
-    public function serverEnvironmentsForClientIpAddresses()
-    {
-        return array(
-            array(array(), '17.172.224.47'),
-            array(array('HTTP_CLIENT_IP' => 'murks'), '17.172.224.47'),
-            array(array('HTTP_CLIENT_IP' => '17.149.160.49'), '17.149.160.49'),
-            array(array('HTTP_CLIENT_IP' => '17.149.160.49', 'HTTP_X_FORWARDED_FOR' => '123.123.123.123'), '17.149.160.49'),
-            array(array('HTTP_X_FORWARDED_FOR' => '123.123.123.123'), '123.123.123.123'),
-            array(array('HTTP_X_FORWARDED_FOR' => '123.123.123.123', 'HTTP_X_FORWARDED' => '209.85.148.101'), '123.123.123.123'),
-            array(array('HTTP_X_FORWARDED_FOR' => '123.123.123', 'HTTP_FORWARDED_FOR' => '209.85.148.101'), '209.85.148.101'),
-            array(array('HTTP_X_FORWARDED_FOR' => '192.168.178.1', 'HTTP_FORWARDED_FOR' => '209.85.148.101'), '209.85.148.101'),
-            array(array('HTTP_X_FORWARDED_FOR' => '123.123.123.123, 209.85.148.101, 209.85.148.102'), '123.123.123.123'),
-            array(array('HTTP_X_CLUSTER_CLIENT_IP' => '209.85.148.101, 209.85.148.102'), '209.85.148.101'),
-            array(array('HTTP_FORWARDED_FOR' => '209.85.148.101'), '209.85.148.101'),
-            array(array('HTTP_FORWARDED' => '209.85.148.101'), '209.85.148.101'),
-            array(array('REMOTE_ADDR' => '127.0.0.1'), '127.0.0.1'),
-        );
-    }
-
-    /**
-     * @test
-     * @dataProvider serverEnvironmentsForClientIpAddresses
-     */
-    public function getClientIpAddressReturnsTheIpAddressDerivedFromSeveralServerEnvironmentVariables(array $serverEnvironment, $expectedIpAddress)
-    {
-        $defaultServerEnvironment = array(
-            'HTTP_USER_AGENT' => 'Flow/' . FLOW_VERSION_BRANCH . '.x',
-            'HTTP_HOST' => 'flow.typo3.org',
-            'SERVER_NAME' => 'typo3.org',
-            'SERVER_ADDR' => '217.29.36.55',
-            'SERVER_PORT' => 80,
-            'REMOTE_ADDR' => '17.172.224.47',
-            'SCRIPT_FILENAME' => FLOW_PATH_WEB . 'index.php',
-            'SERVER_PROTOCOL' => 'HTTP/1.1',
-            'SCRIPT_NAME' => '/index.php',
-            'PHP_SELF' => '/index.php',
-        );
-
-        $request = Request::create(new Uri('http://flow.typo3.org'), 'GET', array(), array(), array_replace($defaultServerEnvironment, $serverEnvironment));
-        $this->assertSame($expectedIpAddress, $request->getClientIpAddress());
-    }
-
-    /**
-     * Data Provider
-     */
     public function acceptHeaderValuesAndCorrespondingListOfMediaTypes()
     {
         return array(
@@ -685,32 +640,6 @@ class RequestTest extends UnitTestCase
     }
 
     /**
-     * RFC 2616 / 14.23 (Host)
-     * @test
-     * @backupGlobals disabled
-     */
-    public function portInProxyHeaderIsAcknowledged()
-    {
-        $scriptName = $_SERVER['SCRIPT_NAME'];
-
-        $_SERVER = array(
-            'HTTP_HOST' => 'dev.blog.rob',
-            'HTTP_X_FORWARDED_PORT' => 2727,
-            'SERVER_NAME' => 'dev.blog.rob',
-            'SERVER_ADDR' => '127.0.0.1',
-            'SERVER_PORT' => '80',
-            'REMOTE_ADDR' => '127.0.0.1',
-            'REQUEST_URI' => '/posts/2011/11/28/laboriosam-soluta-est-minus-molestiae?getKey1=getValue1&getKey2=getValue2',
-            'REQUEST_TIME' => 1326472534
-        );
-
-        $request = Request::create(new Uri('https://dev.blog.rob/foo/bar?baz=quux&coffee=due'), array(), array(), array(), $_SERVER);
-        $this->assertSame(2727, $request->getPort());
-
-        $_SERVER['SCRIPT_NAME'] = $scriptName;
-    }
-
-    /**
      * @test
      */
     public function setContentAlsoAcceptsAFileHandleAsInput()
@@ -736,288 +665,6 @@ class RequestTest extends UnitTestCase
         $this->assertSame($streamHandler, $request->getContent());
         $this->assertEquals('application/octet-stream', $request->getHeader('Content-Type'));
         $this->assertEquals(filesize(__FILE__), $request->getHeader('Content-Length'));
-    }
-
-    /**
-     * @test
-     */
-    public function isSecureReturnsTrueEvenIfTheSchemeIsHttpButTheRequestWasForwardedAndOriginallyWasHttps()
-    {
-        $server = array(
-            'HTTP_X_FORWARDED_PROTO' => 'https',
-            'HTTP_X_FORWARDED_PORT' => '443',
-        );
-
-        $request = Request::create(new Uri('http://acme.com'), 'GET', array(), array(), $server);
-        $this->assertEquals('https://acme.com', (string)$request->getUri());
-        $this->assertEquals('https', $request->getUri()->getScheme());
-        $this->assertTrue($request->isSecure());
-    }
-
-    /**
-     * @test
-     */
-    public function isSecureReturnsFalseIfTheRequestWasForwardedAndOriginallyWasHttp()
-    {
-        $server = array(
-            'HTTP_X_FORWARDED_PROTO' => 'http',
-            'HTTP_X_FORWARDED_PORT' => '80',
-        );
-
-        $request = Request::create(new Uri('https://acme.com'), 'GET', array(), array(), $server);
-        $this->assertEquals('http://acme.com', (string)$request->getUri());
-        $this->assertEquals('http', $request->getUri()->getScheme());
-        $this->assertFalse($request->isSecure());
-    }
-
-    /**
-     * @test
-     */
-    public function isFromTrustedProxyByDefault()
-    {
-        $request = Request::create(new Uri('https://acme.com'), 'GET');
-        $this->assertTrue($request->isFromTrustedProxy());
-    }
-
-    /**
-     * @test
-     */
-    public function isFromTrustedProxyIfRemoteAddressMatchesRange()
-    {
-        $this->trustedProxiesSettings->setValue(['proxies' => ['127.0.0.0/24']]);
-        $request = Request::create(new Uri('https://acme.com'), 'GET');
-        $this->assertTrue($request->isFromTrustedProxy());
-    }
-
-    /**
-     * @test
-     */
-    public function isNotFromTrustedProxyIfNoProxiesAreTrusted()
-    {
-        $this->trustedProxiesSettings->setValue(['proxies' => []]);
-        $request = Request::create(new Uri('https://acme.com'), 'GET');
-        $this->assertFalse($request->isFromTrustedProxy());
-    }
-
-    /**
-     * @test
-     */
-    public function isNotFromTrustedProxyIfRemoteAddressDoesntMatch()
-    {
-        $this->trustedProxiesSettings->setValue(['proxies' => ['10.0.0.1/24']]);
-        $request = Request::create(new Uri('https://acme.com'), 'GET');
-        $this->assertFalse($request->isFromTrustedProxy());
-    }
-
-    /**
-     * @test
-     */
-    public function trustedClientIpAddressIsRemoteAddressIfNoProxiesAreTrusted()
-    {
-        $this->trustedProxiesSettings->setValue(['proxies' => [], 'headers' => [Request::HEADER_CLIENT_IP => 'X-Forwarded-For']]);
-        $request = Request::create(new Uri('https://acme.com'), 'GET', array(), array(), array('HTTP_X_FORWARDED_FOR' => '10.0.0.1'));
-        $this->assertEquals('127.0.0.1', $request->getTrustedClientIpAddress());
-    }
-
-    /**
-     * @test
-     */
-    public function trustedClientIpAddressIsRemoteAddressIfHeaderNotTrusted()
-    {
-        $this->trustedProxiesSettings->setValue(['proxies' => ['127.0.0.1'], 'headers' => []]);
-        $request = Request::create(new Uri('https://acme.com'), 'GET', array(), array(), array('HTTP_X_FORWARDED_FOR' => '10.0.0.1'));
-        $this->assertEquals('127.0.0.1', $request->getTrustedClientIpAddress());
-    }
-
-    /**
-     * @test
-     */
-    public function trustedClientIpAddressIsForwardedForAddressIfProxyTrusted()
-    {
-        $this->trustedProxiesSettings->setValue(['proxies' => ['127.0.0.1'], 'headers' => [Request::HEADER_CLIENT_IP => 'X-Forwarded-For']]);
-        $request = Request::create(new Uri('https://acme.com'), 'GET', array(), array(), array('HTTP_X_FORWARDED_FOR' => '10.0.0.1'));
-        $this->assertEquals('10.0.0.1', $request->getTrustedClientIpAddress());
-    }
-
-    /**
-     * @test
-     */
-    public function trustedClientIpAddressIsFirstForwardedForAddressIfAllProxiesTrusted()
-    {
-        $this->trustedProxiesSettings->setValue(['proxies' => '*', 'headers' => [Request::HEADER_CLIENT_IP => 'X-Forwarded-For']]);
-        $request = Request::create(new Uri('https://acme.com'), 'GET', array(), array(), array('HTTP_X_FORWARDED_FOR' => '10.0.0.1, 10.0.0.2, 10.0.0.3'));
-        $this->assertEquals('10.0.0.1', $request->getTrustedClientIpAddress());
-    }
-
-    /**
-     * @test
-     */
-    public function trustedClientIpAddressIsRightMostForwardedForAddressThatIsNotTrusted()
-    {
-        $this->trustedProxiesSettings->setValue(['proxies' => ['127.0.0.1','10.0.0.1/24'], 'headers' => [Request::HEADER_CLIENT_IP => 'X-Forwarded-For']]);
-        $request = Request::create(new Uri('https://acme.com'), 'GET', array(), array(), array('HTTP_X_FORWARDED_FOR' => '198.155.23.17, 215.0.0.1, 10.0.0.1, 10.0.0.2'));
-        $this->assertEquals('215.0.0.1', $request->getTrustedClientIpAddress());
-    }
-
-    /**
-     * @test
-     */
-    public function trustedClientIpAddressIsRemoteAddressIfTheHeaderIsNotTrusted()
-    {
-        $this->trustedProxiesSettings->setValue(['proxies' => '*', 'headers' => [Request::HEADER_CLIENT_IP => 'X-Forwarded-Ip']]);
-        $request = Request::create(new Uri('https://acme.com'), 'GET', array(), array(), array('HTTP_X_FORWARDED_FOR' => '10.0.0.1'));
-        $this->assertEquals('127.0.0.1', $request->getTrustedClientIpAddress());
-    }
-
-    /**
-     * @test
-     */
-    public function portIsNotOverridenIfTheHeaderIsNotTrusted()
-    {
-        $this->trustedProxiesSettings->setValue(['proxies' => '*', 'headers' => []]);
-        $request = Request::create(new Uri('http://acme.com'), 'GET', array(), array(), array('HTTP_X_FORWARDED_PORT' => '443'));
-        $this->assertEquals(80, $request->getPort());
-    }
-
-    /**
-     * @test
-     */
-    public function protocolIsNotOverridenIfTheHeaderIsNotTrusted()
-    {
-        $this->trustedProxiesSettings->setValue(['proxies' => '*', 'headers' => []]);
-        $request = Request::create(new Uri('http://acme.com'), 'GET', array(), array(), array('HTTP_X_FORWARDED_PROTO' => 'https'));
-        $this->assertEquals('http', $request->getUri()->getScheme());
-    }
-
-    /**
-     * @test
-     */
-    public function hostIsNotOverridenIfTheHeaderIsNotTrusted()
-    {
-        $this->trustedProxiesSettings->setValue(['proxies' => '*', 'headers' => []]);
-        $request = Request::create(new Uri('http://acme.com'), 'GET', array(), array(), array('HTTP_X_FORWARDED_HOST' => 'neos.io'));
-        $this->assertEquals('acme.com', $request->getUri()->getHost());
-    }
-    /**
-     * @return array
-     */
-    public function forwardHeaderTestsDataProvider()
-    {
-        return array(
-            array(
-                'forwardedProtocol' => null,
-                'forwardedPort' => null,
-                'requestUri' => 'http://acme.com',
-                'expectedUri' => 'http://acme.com',
-            ),
-
-            // forwarded protocol overrules requested protocol
-            array(
-                'forwardedProtocol' => 'https',
-                'forwardedPort' => null,
-                'requestUri' => 'http://acme.com',
-                'expectedUri' => 'https://acme.com',
-            ),
-            array(
-                'forwardedProtocol' => 'https',
-                'forwardedPort' => null,
-                'requestUri' => 'https://acme.com',
-                'expectedUri' => 'https://acme.com',
-            ),
-            array(
-                'forwardedProtocol' => 'http',
-                'forwardedPort' => null,
-                'requestUri' => 'https://acme.com',
-                'expectedUri' => 'http://acme.com',
-            ),
-            array(
-                'forwardedProtocol' => 'http',
-                'forwardedPort' => null,
-                'requestUri' => 'http://acme.com',
-                'expectedUri' => 'http://acme.com',
-            ),
-
-            // forwarded port overrules requested port
-            array(
-                'forwardedProtocol' => null,
-                'forwardedPort' => 80,
-                'requestUri' => 'http://acme.com',
-                'expectedUri' => 'http://acme.com',
-            ),
-            array(
-                'forwardedProtocol' => null,
-                'forwardedPort' => '8080',
-                'requestUri' => 'http://acme.com',
-                'expectedUri' => 'http://acme.com:8080',
-            ),
-            array(
-                'forwardedProtocol' => null,
-                'forwardedPort' => 8080,
-                'requestUri' => 'http://acme.com:8000',
-                'expectedUri' => 'http://acme.com:8080',
-            ),
-            array(
-                'forwardedProtocol' => null,
-                'forwardedPort' => '443',
-                'requestUri' => 'https://acme.com',
-                'expectedUri' => 'https://acme.com',
-            ),
-
-            // forwarded protocol & port
-            array(
-                'forwardedProtocol' => 'http',
-                'forwardedPort' => 80,
-                'requestUri' => 'http://acme.com',
-                'expectedUri' => 'http://acme.com',
-            ),
-            array(
-                'forwardedProtocol' => 'http',
-                'forwardedPort' => 8080,
-                'requestUri' => 'http://acme.com',
-                'expectedUri' => 'http://acme.com:8080',
-            ),
-            array(
-                'forwardedProtocol' => 'http',
-                'forwardedPort' => 443,
-                'requestUri' => 'https://acme.com',
-                'expectedUri' => 'http://acme.com:443',
-            ),
-            array(
-                'forwardedProtocol' => 'https',
-                'forwardedPort' => 443,
-                'requestUri' => 'http://acme.com',
-                'expectedUri' => 'https://acme.com',
-            ),
-            array(
-                'forwardedProtocol' => 'https',
-                'forwardedPort' => 443,
-                'requestUri' => 'https://acme.com',
-                'expectedUri' => 'https://acme.com',
-            ),
-            array(
-                'forwardedProtocol' => 'https',
-                'forwardedPort' => 80,
-                'requestUri' => 'https://acme.com',
-                'expectedUri' => 'https://acme.com:80',
-            ),
-        );
-    }
-
-    /**
-     * @test
-     * @dataProvider forwardHeaderTestsDataProvider
-     */
-    public function forwardHeaderTests($forwardedProtocol, $forwardedPort, $requestUri, $expectedUri)
-    {
-        $server = array();
-        if ($forwardedProtocol !== null) {
-            $server['HTTP_X_FORWARDED_PROTO'] = $forwardedProtocol;
-        }
-        if ($forwardedPort !== null) {
-            $server['HTTP_X_FORWARDED_PORT'] = $forwardedPort;
-        }
-        $request = Request::create(new Uri($requestUri), 'GET', array(), array(), $server);
-        $this->assertEquals($expectedUri, (string)$request->getUri());
     }
 
     /**
